+++ RELEASE_NOTES.md
@@ -3,7 +3,7 @@ Release Notes
 
 ## vNext
 * Load Balancer: Generate an IPv6 public IP address using the `ip_v6` option.
-* Virtual Machine Scale Sets: Support for VM Scale Sets.
+* Virtual Machine Scale Sets: Support for VM Scale Sets. for 
 
 ## 1.7.31
 * Subnets: Support for `depends_on` when defining standalone subnets.